refactor(api): rename /private/config/google_oauth to /private/config/builder
Generalizes the route name so it can serve other build-onboarding
config in the future without another rename. The response shape and
GOOGLE_OAUTH_* env var names stay as-is — the values it serves today
are still Google "Desktop app" OAuth credentials for the CLI's
`build init` Android flow. File and import variable renamed in
lockstep:
  config_google_oauth.ts -> config_builder.ts
  configGoogleOauth      -> configBuilder

Both router entry points updated:
  - cloudflare_workers/api/index.ts
  - supabase/functions/private/index.ts
